\name{opticut-package}
\alias{opticut-package}
\alias{opticut}
\docType{package}
\title{
\packageTitle{opticut}
}
\description{
\packageDescription{opticut}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{opticut}
\packageIndices{opticut}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{opticut}

Maintainer: \packageMaintainer{opticut}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
